+++ test/test_source_index.mli
@@ -1,0 +1,4 @@
+(** Test suite for {!Cascade.Source_index}. *)
+
+val suite : string * unit Alcotest.test_case list
+(** [suite] test suite. *)